test request input

+++ CafeteriaApp/CafeteriaApp.Backend/Requests/Admin.php
@@ -8,20 +8,20 @@ require_once ('CheckResult.php');
 if ($_SERVER["REQUEST_METHOD"] == "POST")
 {
 	$data = json_decode(file_get_contents("php://input"));
-	addAdmin($conn,$data->UserId);
+	if ($data->UserId !== null && is_int($data->UserId))
+	{
+		addAdmin($conn,$data->UserId);
+	}
 }
 
 if ($_SERVER["REQUEST_METHOD"] == "DELETE")
 {
-	if (isset($_GET["userId"]) && $_GET["userId"] != null)
+	if (isset($_GET["userId"]) && $_GET["userId"] !== null && is_int($_GET["userId"]))
 	{
 		deleteAdminByUserId($conn,$_GET["userId"]);
 	}
-	else if (isset($_GET["adminId"]) && $_GET["adminId"] != null)
-	{
-		deleteAdmin($conn,$_GET["adminId"]);
-	}
 }
+
 require_once("CafeteriaApp.Backend/footer.php");
 
 ?>
